refactor: move breed map to utils
