+++ server.js
@@ -6,33 +6,7 @@ import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
 import morgan from 'morgan';
 import fetchCleanPhoto from './utils/fetchCleanPhoto.js';
-
-// Map Afrikaans breed names to their English equivalents so Unsplash can
-// return matching results. The search terms are almost entirely in English,
-// so Afrikaans queries often yield a 404 response.
-export const breedMap = {
-  // Dogs
-  'Afgaanse hond': 'Afghan hound',
-  'Goudkleurige Apporteerhond': 'Golden Retriever',
-  'Dalmatiese hond': 'Dalmatian',
-  Dashond: 'Dachshund',
-  'Franse Dashond': 'Basset Hound',
-  Jagwindhond: 'greyhound',
-  Bulhond: 'Bulldog',
-  'Engelse Patryshond': 'English pointer',
-  Bloedhond: 'Bloodhound',
-  'Rhodesiese Rifrughond': 'Rhodesian Ridgeback',
-  // Cats (week 2)
-  'Rob bruin Himalajan kat': 'Himalayan cat',
-  'Siamese kat': 'Siamese cat',
-  'Turkse Van kat': 'Turkish Van cat',
-  'Chinchilla kat': 'Chinchilla cat',
-  'Rooi Persiese kat': 'Red Persian cat',
-  'Rokerige Swart Persiese kat': 'Smoky black Persian cat',
-  'Turkse Angora kat': 'Turkish Angora cat',
-  'Abbesini\u00e9r kat': 'Abyssinian cat',
-  'Ragdoll kat': 'Ragdoll cat',
-};
+import { breedMap } from './utils/breedMap.js';
 
 dotenv.config();
 
@@ -137,4 +111,5 @@ if (process.env.NODE_ENV !== 'test') {
   });
 }
 
+export { breedMap };
 export default app;
